fix(sidebar): modify menu wrapper behavior
