+++ core/src/sidebar/heading/heading.tsx
@@ -1,6 +1,4 @@
-import s from "./heading.module.css"
+import s from "./heading.module.css";
 export const Heading = (props: { children: string }): JSX.Element => {
-	return <div className={s.heading}>
-		{props.children}
-	</div>;
+	return <div className={s.heading}>{props.children}</div>;
 };
