+++ core/src/sidebar/sidebar.module.css
@@ -1,3 +1,8 @@
+.stories {
+	display: flex;
+	width: 100%;
+}
+
 .container {
 	width: 232px;
 	background-color: white;
